Revert "eslintのプラグインをインストール"
