+++ .env
@@ -1,2 +1,2 @@
-# NEXT_PUBLIC_BACKEND_URL=http://localhost
-NEXT_PUBLIC_BACKEND_URL=https://api.graduation-front.com/project
+NEXT_PUBLIC_BACKEND_URL=http://localhost
+# NEXT_PUBLIC_BACKEND_URL=https://api.graduation-front.com/project
