+++ package.json
@@ -17,15 +17,12 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.12.1",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
     "@next/eslint-plugin-next": "^10.0.4",
     "@tailwindcss/forms": "^0.5.2",
     "autoprefixer": "^10.4.2",
     "eslint": "^8.27.0",
     "eslint-config-next": "13.0.3",
     "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-jsx": "^0.1.0",
     "eslint-plugin-prettier": "^3.3.0",
     "postcss": "^8.4.6",
     "prettier": "2.2.1",
