+++ src/components/Layouts/AppLayout.js
@@ -1,6 +1,7 @@
 import Navigation from '@/components/Layouts/Navigation'
 
-const AppLayout = ({ user, children }) => {
+const AppLayout = ({user, children }) => {
+
     return (
         <div className="min-h-screen bg-gray-100">
             <Navigation user={user} />
