+++ src/components/Layouts/Layout.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-const Layout = ({ children }) => {
+const Layout = ({children}) => {
     return (
         <div className="min-h-screen bg-gray-100">
             {/* Page Content */}
