+++ src/components/Layouts/Navigation.js
@@ -1,14 +1,29 @@
 import Dropdown from '@/components/Dropdown'
-import { ResponsiveNavButton } from '@/components/ResponsiveNavLink'
+import Link from 'next/link'
+import NavLink from '@/components/NavLink'
+import ResponsiveNavLink, {
+    ResponsiveNavButton,
+} from '@/components/ResponsiveNavLink'
 import { DropdownButton } from '@/components/DropdownLink'
 import { useAuth } from '@/hooks/auth'
-import { useState } from 'react'
+import { useRouter } from 'next/router'
+import { use, useState } from 'react'
 
 const Navigation = ({ user }) => {
+    const router = useRouter()
+
     const { logout } = useAuth()
 
     // ハンバーガーメニューが開いているかどうか
     const [open, setOpen] = useState(false)
+
+    if (user) {
+        const path = `${process.env.NEXT_PUBLIC_BACKEND_URL}/${user.icon_path}`
+    } else {
+        const path = 'icon.png'
+    }
+    // console.log(user.icon_path);
+    // const icon = user.icon_path;
 
     return (
         <nav className="bg-white border-b border-gray-100">
