+++ src/hooks/auth.js
@@ -91,7 +91,7 @@ export const useAuth = ({ middleware, redirectIfAuthenticated } = {}) => {
     }
 
     const logout = async () => {
-        if (!error) {
+        if (! error) {
             await axios.post('/logout').then(() => mutate())
         }
