+++ src/lib/axios.js
@@ -4,7 +4,7 @@ const axios = Axios.create({
     baseURL: process.env.NEXT_PUBLIC_BACKEND_URL,
     headers: {
         'X-Requested-With': 'XMLHttpRequest',
-        Accept: 'application/json',
+        'Accept': 'application/json',
     },
     withCredentials: true,
 })
