+++ src/pages/404.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 const NotFoundPage = () => (
     <div className="relative flex items-top justify-center min-h-screen bg-gray-100 dark:bg-gray-900 sm:items-center sm:pt-0">
         <div className="max-w-xl mx-auto sm:px-6 lg:px-8">
