+++ src/pages/forgot-password.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import ApplicationLogo from '@/components/ApplicationLogo'
 import AuthCard from '@/components/AuthCard'
 import AuthSessionStatus from '@/components/AuthSessionStatus'
 import Button from '@/components/Button'
@@ -6,6 +6,7 @@ import GuestLayout from '@/components/Layouts/GuestLayout'
 import Input from '@/components/Input'
 import InputError from '@/components/InputError'
 import Label from '@/components/Label'
+import Link from 'next/link'
 import { useAuth } from '@/hooks/auth'
 import { useState } from 'react'
