+++ src/pages/indexDefault.js
@@ -1,8 +1,8 @@
-import React from 'react'
 import styles from '@/styles/index.module.css'
 import Head from 'next/head'
 import Link from 'next/link'
 import { useAuth } from '@/hooks/auth'
+import axios from '@/lib/axios'
 
 export default function Home() {
     const { user } = useAuth({ middleware: 'guest' })
@@ -16,7 +16,9 @@ export default function Home() {
             <div className={styles.container}>
                 <div className={styles.buttonContainer}>
                     {user ? (
-                        <Link href="/home" className={styles.button}>
+                        <Link
+                            href="/home"
+                            className={styles.button}>
                             Home
                         </Link>
                     ) : (
@@ -27,13 +29,21 @@ export default function Home() {
                                 Login
                             </Link>
 
-                            <Link href="/register" className={styles.button}>
+                            <Link
+                                href="/register"
+                                className={styles.button}>
                                 Register
                             </Link>
                         </>
                     )}
                 </div>
             </div>
+            
+            <button onClick={() => {
+                axios.get('/api/hoge').then((res) => {
+                    console.log(res);
+                })
+            }}>ボタン</button>
         </>
     )
 }
