+++ src/pages/login.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import ApplicationLogo from '@/components/ApplicationLogo'
 import AuthCard from '@/components/AuthCard'
 import AuthSessionStatus from '@/components/AuthSessionStatus'
@@ -12,41 +11,43 @@ import { useAuth } from '@/hooks/auth'
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 
-const PasswordReset = () => {
+const Login = () => {
     const router = useRouter()
 
-    const { resetPassword } = useAuth({ middleware: 'guest' })
+    const { login } = useAuth({
+        middleware: 'guest',
+        redirectIfAuthenticated: '/',
+    })
 
     const [email, setEmail] = useState('')
     const [password, setPassword] = useState('')
-    const [passwordConfirmation, setPasswordConfirmation] = useState('')
+    const [shouldRemember, setShouldRemember] = useState(false)
     const [errors, setErrors] = useState([])
     const [status, setStatus] = useState(null)
 
-    const submitForm = event => {
+    useEffect(() => {
+        if (router.query.reset?.length > 0 && errors.length === 0) {
+            setStatus(atob(router.query.reset))
+        } else {
+            setStatus(null)
+        }
+    })
+
+    const submitForm = async event => {
         event.preventDefault()
 
-        resetPassword({
+        login({
             email,
             password,
-            password_confirmation: passwordConfirmation,
+            remember: shouldRemember,
             setErrors,
             setStatus,
         })
     }
 
-    useEffect(() => {
-        setEmail(router.query.email || '')
-    }, [router.query.email])
-
     return (
         <GuestLayout>
-            <AuthCard
-                logo={
-                    <Link href="/">
-                        <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
-                    </Link>
-                }>
+            <AuthCard>
                 {/* Session Status */}
                 <AuthSessionStatus className="mb-4" status={status} />
 
@@ -71,6 +72,7 @@ const PasswordReset = () => {
                     {/* Password */}
                     <div className="mt-4">
                         <Label htmlFor="password">Password</Label>
+
                         <Input
                             id="password"
                             type="password"
@@ -78,6 +80,7 @@ const PasswordReset = () => {
                             className="block mt-1 w-full"
                             onChange={event => setPassword(event.target.value)}
                             required
+                            autoComplete="current-password"
                         />
 
                         <InputError
@@ -86,31 +89,39 @@ const PasswordReset = () => {
                         />
                     </div>
 
-                    {/* Confirm Password */}
-                    <div className="mt-4">
-                        <Label htmlFor="passwordConfirmation">
-                            Confirm Password
-                        </Label>
+                    {/* Remember Me */}
+                    <div className="block mt-4">
+                        <label
+                            htmlFor="remember_me"
+                            className="inline-flex items-center">
+                            <input
+                                id="remember_me"
+                                type="checkbox"
+                                name="remember"
+                                className="rounded border-gray-300 text-indigo-600 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
+                                onChange={event =>
+                                    setShouldRemember(event.target.checked)
+                                }
+                            />
 
-                        <Input
-                            id="passwordConfirmation"
-                            type="password"
-                            value={passwordConfirmation}
-                            className="block mt-1 w-full"
-                            onChange={event =>
-                                setPasswordConfirmation(event.target.value)
-                            }
-                            required
-                        />
-
-                        <InputError
-                            messages={errors.password_confirmation}
-                            className="mt-2"
-                        />
+                            <span className="ml-2 text-sm text-gray-600">
+                                Remember me
+                            </span>
+                        </label>
                     </div>
 
-                    <div className="flex items-center justify-end mt-4">
-                        <Button>Reset Password</Button>
+                    <div className="flex items-center justify-center mt-4">
+                        <Link
+                            href="/register"
+                            className="underline text-sm text-gray-600 hover:text-gray-900">
+                            新規登録の方
+                        </Link>
+                        <Link
+                            href="/forgot-password"
+                            className="underline text-sm text-gray-600 hover:text-gray-900 ml-3">
+                            パスワードを忘れた方
+                        </Link>
+                        <Button className="ml-3">Login</Button>
                     </div>
                 </form>
             </AuthCard>
@@ -118,4 +129,4 @@ const PasswordReset = () => {
     )
 }
 
-export default PasswordReset
+export default Login
