+++ src/pages/password-reset/[token].js
@@ -1,4 +1,4 @@
-import React from 'react'
+import ApplicationLogo from '@/components/ApplicationLogo'
 import AuthCard from '@/components/AuthCard'
 import AuthSessionStatus from '@/components/AuthSessionStatus'
 import Button from '@/components/Button'
@@ -11,43 +11,41 @@ import { useAuth } from '@/hooks/auth'
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 
-const Login = () => {
+const PasswordReset = () => {
     const router = useRouter()
 
-    const { login } = useAuth({
-        middleware: 'guest',
-        redirectIfAuthenticated: '/',
-    })
+    const { resetPassword } = useAuth({ middleware: 'guest' })
 
     const [email, setEmail] = useState('')
     const [password, setPassword] = useState('')
-    const [shouldRemember, setShouldRemember] = useState(false)
+    const [passwordConfirmation, setPasswordConfirmation] = useState('')
     const [errors, setErrors] = useState([])
     const [status, setStatus] = useState(null)
 
-    useEffect(() => {
-        if (router.query.reset?.length > 0 && errors.length === 0) {
-            setStatus(atob(router.query.reset))
-        } else {
-            setStatus(null)
-        }
-    })
-
-    const submitForm = async event => {
+    const submitForm = event => {
         event.preventDefault()
 
-        login({
+        resetPassword({
             email,
             password,
-            remember: shouldRemember,
+            password_confirmation: passwordConfirmation,
             setErrors,
             setStatus,
         })
     }
 
+    useEffect(() => {
+        setEmail(router.query.email || '')
+    }, [router.query.email])
+
     return (
         <GuestLayout>
-            <AuthCard>
+            <AuthCard
+                logo={
+                    <Link href="/">
+                        <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
+                    </Link>
+                }>
                 {/* Session Status */}
                 <AuthSessionStatus className="mb-4" status={status} />
 
@@ -72,7 +70,6 @@ const Login = () => {
                     {/* Password */}
                     <div className="mt-4">
                         <Label htmlFor="password">Password</Label>
-
                         <Input
                             id="password"
                             type="password"
@@ -80,7 +77,6 @@ const Login = () => {
                             className="block mt-1 w-full"
                             onChange={event => setPassword(event.target.value)}
                             required
-                            autoComplete="current-password"
                         />
 
                         <InputError
@@ -89,39 +85,31 @@ const Login = () => {
                         />
                     </div>
 
-                    {/* Remember Me */}
-                    <div className="block mt-4">
-                        <label
-                            htmlFor="remember_me"
-                            className="inline-flex items-center">
-                            <input
-                                id="remember_me"
-                                type="checkbox"
-                                name="remember"
-                                className="rounded border-gray-300 text-indigo-600 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
-                                onChange={event =>
-                                    setShouldRemember(event.target.checked)
-                                }
-                            />
+                    {/* Confirm Password */}
+                    <div className="mt-4">
+                        <Label htmlFor="passwordConfirmation">
+                            Confirm Password
+                        </Label>
 
-                            <span className="ml-2 text-sm text-gray-600">
-                                Remember me
-                            </span>
-                        </label>
+                        <Input
+                            id="passwordConfirmation"
+                            type="password"
+                            value={passwordConfirmation}
+                            className="block mt-1 w-full"
+                            onChange={event =>
+                                setPasswordConfirmation(event.target.value)
+                            }
+                            required
+                        />
+
+                        <InputError
+                            messages={errors.password_confirmation}
+                            className="mt-2"
+                        />
                     </div>
 
-                    <div className="flex items-center justify-center mt-4">
-                        <Link
-                            href="/register"
-                            className="underline text-sm text-gray-600 hover:text-gray-900">
-                            新規登録の方
-                        </Link>
-                        <Link
-                            href="/forgot-password"
-                            className="underline text-sm text-gray-600 hover:text-gray-900 ml-3">
-                            パスワードを忘れた方
-                        </Link>
-                        <Button className="ml-3">Login</Button>
+                    <div className="flex items-center justify-end mt-4">
+                        <Button>Reset Password</Button>
                     </div>
                 </form>
             </AuthCard>
@@ -129,4 +117,4 @@ const Login = () => {
     )
 }
 
-export default Login
+export default PasswordReset
