+++ src/styles/profile.module.css
@@ -39,4 +39,5 @@
 .userInfo {
   display: flex;
   flex-direction: column;
+  align-items: start;
 }
